Moving functions to calendar-aux and bug fixes
